DEC-244: Broke out the custom generators into its own space and files to separate the app specific generators from the generic.
