+++ app/controllers/admin/administrators_controller.rb
@@ -5,7 +5,7 @@ module Admin
       administrators = AdministratorQuery.new.list
       @administrators = AdministratorListDecorator.new(administrators)
 
-      cache_key = CacheKeys::Generator.new(key_generator: CacheKeys::Generator::Administrators,
+      cache_key = CacheKeys::Generator.new(key_generator: CacheKeys::Custom::Administrators,
                                            action: "index",
                                            decorated_administrators: @administrators)
       fresh_when(etag: cache_key.generate)
