+++ app/controllers/v1/collections_controller.rb
@@ -3,7 +3,7 @@ module V1
     def index
       @collections = CollectionQuery.new.public_collections
 
-      cache_key = CacheKeys::Generator.new(key_generator: CacheKeys::Generator::V1Collections,
+      cache_key = CacheKeys::Generator.new(key_generator: CacheKeys::Custom::V1Collections,
                                            action: "index",
                                            collections: @collections)
       fresh_when(etag: cache_key.generate)
@@ -12,7 +12,7 @@ module V1
     def show
       @collection = CollectionQuery.new.public_find(params[:id])
 
-      cache_key = CacheKeys::Generator.new(key_generator: CacheKeys::Generator::V1Collections,
+      cache_key = CacheKeys::Generator.new(key_generator: CacheKeys::Custom::V1Collections,
                                            action: "show",
                                            collection: @collection)
       fresh_when(etag: cache_key.generate)
