+++ app/controllers/v1/showcases_controller.rb
@@ -4,7 +4,7 @@ module V1
       collection = CollectionQuery.new.public_find(params[:collection_id])
       @collection = CollectionJSONDecorator.new(collection)
 
-      cache_key = CacheKeys::Generator.new(key_generator: CacheKeys::Generator::V1Showcases,
+      cache_key = CacheKeys::Generator.new(key_generator: CacheKeys::Custom::V1Showcases,
                                            action: "index",
                                            collection: collection)
       fresh_when(etag: cache_key.generate)
@@ -14,7 +14,7 @@ module V1
       showcase = ShowcaseQuery.new.public_find(params[:id])
       @showcase = ShowcaseJSONDecorator.new(showcase)
 
-      cache_key = CacheKeys::Generator.new(key_generator: CacheKeys::Generator::V1Showcases,
+      cache_key = CacheKeys::Generator.new(key_generator: CacheKeys::Custom::V1Showcases,
                                            action: "show",
                                            showcase: showcase)
       fresh_when(etag: cache_key.generate)
