+++ app/values/cache_keys/custom/administrators.rb
@@ -1,0 +1,10 @@
+module CacheKeys
+  module Custom
+    # Generator for admin/administrators_controller
+    class Administrators
+      def index(decorated_administrators:)
+        CacheKeys::DecoratedActiveRecord.new.generate(record: decorated_administrators)
+      end
+    end
+  end
+end
