+++ app/values/cache_keys/custom/v1_items.rb
@@ -1,0 +1,14 @@
+module CacheKeys
+  module Custom
+    # Generator for v1/items_controller
+    class V1Items
+      def index(collection:)
+        CacheKeys::ActiveRecord.new.generate(record: [collection, collection.items])
+      end
+
+      def show(item:)
+        CacheKeys::ActiveRecord.new.generate(record: [item, item.collection, item.children])
+      end
+    end
+  end
+end
