+++ app/values/cache_keys/custom/v1_showcases.rb
@@ -1,0 +1,14 @@
+module CacheKeys
+  module Custom
+    # Generator for v1/showcases_controller
+    class V1Showcases
+      def index(collection:)
+        CacheKeys::ActiveRecord.new.generate(record: [collection, collection.showcases])
+      end
+
+      def show(showcase:)
+        CacheKeys::ActiveRecord.new.generate(record: [showcase, showcase.collection, showcase.sections, showcase.items])
+      end
+    end
+  end
+end
